Some update And Start result

+++ database/migrations/2018_08_19_005553_create_clsrol_table.php
@@ -15,7 +15,7 @@ class CreateClsrolTable extends Migration
     {
         Schema::create('clsrol', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('stdid');
+            $table->unsignedBigInteger('stdid');
             $table->string('sclcd', 11);
             $table->unsignedInteger('stdcls');
             $table->unsignedInteger('stdrol');
