+++ database/migrations/2018_09_02_112842_ceate_user_profile_table.php
@@ -15,7 +15,7 @@ class CeateUserProfileTable extends Migration
     {
         Schema::create('usrpro', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('usrid');
+            $table->unsignedBigInteger('usrid');
             $table->string('sclcd', 11);
             $table->string('abt', 255);
             $table->string('fthr', 30);
